setup jest rather than mocha for unit tests  🐿 v2.8.0
