+++ .gitignore
@@ -7,3 +7,4 @@ bower_components/
 /pa11y_screenCapture/*
 *.env*
 /.vscode
+/coverage
