+++ package.json
@@ -35,6 +35,7 @@
     "sinon-chai": "^3.1.0"
   },
   "dependencies": {
+    "jest": "^23.1.0",
     "moment": "^2.22.1"
   }
 }
